refactor : infrastructure 패키지 user 모듈로 이동
