+++ src/main/java/com/meoguri/linkocean/domain/user/infrastructure/oauth/GoogleOAuthToken.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.infrastructure.oauth.google;
+package com.meoguri.linkocean.domain.user.infrastructure.oauth;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
